fix(apps/cave/airdrop): mobile ui
